requirements.txt
fastapi>=0.100.0
pydantic>=2.0
uvicorn
python-dotenv
cryptography